avoid duplication of virtual hosts
before this patch, redirection virtual hosts were created with a duplicate ServerName, which seems to be working, but is needlessly confusing

after this patch, the redirection vhosts are created only based on the aliases provided and not the site_url

this implies that to do generic redirections (e.g. for SSL) you may need to push your site_url in the aliases array. that has been done in SSL so that all hosts (ie. aliases and the site_url) are redirected to the site_url, which is not appropriate for regular vhosts (because it would mean a infinite redirection loop)