did some weird stuff by introducing the parent property, and automatic oid loading.
d($object)->web_server now refers to the object with the key specified in the
--web_server option.

So instead of provision_invoke('http') to access the services, which only allowed
us to access the local (on this server) services, we can now specify external
servers for specific requirements.

Once having set up the @server_WEB_FRONT_1 alias, we will be able to tell the platform
that it is exposed via that web server.

There's still some crazy stuff to work out though.