Server verify .. _WARNING_ this will break your install.
almost all the variables that used to be set in the platform/drushrc.php are now removed.
you need to verify your server first (done by not specifying a --root, or being in a drupal dir when you run the command).

We need to bootstrap the values from hosting somehow too...