add a command to parse --backend output
i got tired of stabbing my eyes with inline JSON output so I wrote a
very very simple command that just parses STDIN as a --backend output
command. this can be useful for debugging when calling backend commands
without thinking of removing --backend, or in my case, to pass data
through STDIN as JSON, but still have a proper output.